Convert Weights to be a float instead of an int
Give mores flexibility

+++ Tetris.Engine.AI/Algorithms/IAlgorithm.cs
@@ -2,6 +2,6 @@
 {
     public interface IAlgorithm
     {
-        int CalculateFitness(bool[][] gameBoard);
+        float CalculateFitness(bool[][] gameBoard);
     }
 }
